Move CLI::Main code to CLI::EnvFileConsolidator
Now, CLI::Main will start to delegate tasks to different classes based
on what the user is trying to do (learned from flags).
